Update code.js
I am unsure if this is the way we are suppose to be doing the reverse insertion sort. But the logic seems to make sense to me since we are starting at the end of the array and working our way to the zero index and ordering the values of the array in descending order.